Put all the playbook tests in the same file
